+++ lib/api.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
+import 'package:jotform/models/form.dart';
 import 'package:jotform/models/submission.dart';
 
 class JotformAPI {
@@ -15,11 +16,37 @@ class JotformAPI {
         queryParameters: {...params, "apiKey": apiKey},
       );
 
+  Stream<JotformForm> getForms({
+    int offset = 0,
+    int limit = 20,
+    String? filter,
+    JotformOrderBy? orderBy,
+  }) async* {
+    yield* await http
+        .get(getUrl("user/forms", params: {
+      "offset": offset.toString(),
+      "limit": limit.toString(),
+      if (filter != null) "filter": filter,
+      if (orderBy != null) "orderBy": orderBy.name.replaceAll("_", "")
+    }))
+        .then((value) {
+      print(value.body);
+
+      Map<String, dynamic> body = jsonDecode(value.body);
+      if (body["limit-left"] is int) {
+        limitLeft = body["limit-left"];
+      }
+      return Stream.fromIterable((body["content"] as List<dynamic>)
+          .whereType<Map<String, dynamic>>()
+          .map((i) => JotformFormMapper.fromMap(i)));
+    });
+  }
+
   Stream<JotformSubmission> getSubmissions(String id,
       {int offset = 0,
       int limit = 20,
       String? filter,
-      JotformSubmissionsOrderBy? orderBy}) async* {
+      JotformOrderBy? orderBy}) async* {
     yield* await http
         .get(getUrl("form/$id/submissions", params: {
       "offset": offset.toString(),
@@ -39,7 +66,7 @@ class JotformAPI {
   }
 }
 
-enum JotformSubmissionsOrderBy {
+enum JotformOrderBy {
   id,
   username,
   title,