+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+* Added user/forms endpoint
+* Added models for forms
+
 ## 1.0.1
 * Added Submission & Answer models and generally cleaned up the json
 * Added basic api with limit tracker